refactor: split pptx text box semantic tests
